+++ lib/teamlab/modules/calendar.rb
@@ -24,6 +24,14 @@ module Teamlab
       @request.get([calendar_id.to_s, 'sharing'])
     end
 
+    def get_event_history_by_id(event_id)
+      @request.get(['events', event_id.to_s, 'historybyid'])
+    end
+
+    def get_event_history_by_uid(event_uid)
+      @request.get(['events', event_uid.to_s, 'historybyuid'])
+    end
+
     def get_icalc_feed(calendar_id, signature)
       @request.get([calendar_id.to_s, 'ical', signature.to_s])
     end
@@ -44,23 +52,25 @@ module Teamlab
       @request.post(%w(calendarurl), { iCalUrl: ical_url, name: name }.merge(options))
     end
 
-    #============== TODO: OPTIONAL VARIABLES
-
     def add_event(calendar_id, name, options = {})
       @request.post([calendar_id.to_s, 'event'], { name: name }.merge(options))
+    end
+
+    def add_icsevent(calendar_id, ics, options = {})
+      @request.post(['icsevent'], { calendarId: calendar_id.to_s, ics: ics.to_s }.merge(options))
     end
 
     def import_ical(calendar_id, file)
       @request.post([calendar_id.to_s, 'import'], somefile: File.new(file))
     end
 
-    #============== TODO: OPTIONAL VARIABLES
+    def import_events(calendar_id, ical)
+      @request.post(['importIcs'], calendarId: calendar_id.to_s, iCalString: ical.to_s)
+    end
 
     def update_calendar(calendar_id, name, time_zone, options = {})
       @request.put([calendar_id.to_s], { name: name, timeZone: time_zone }.merge(options))
     end
-
-    #============== TODO: OPTIONAL VARIABLES
 
     def update_calendar_user_view(calendar_id, name, time_zone, options = {})
       @request.put([calendar_id.to_s, 'view'], { name: name, timeZone: time_zone }.merge(options))
@@ -70,10 +80,12 @@ module Teamlab
       @request.put(%w(subscriptions manage), states: states)
     end
 
-    #============== TODO: OPTIONAL VARIABLES
-
     def update_event(calendar_id, event_id, name, options = {})
       @request.put([calendar_id.to_s, event_id.to_s], { name: name }.merge(options))
+    end
+
+    def update_icsevent(calendar_id, event_id, ics, options = {})
+      @request.put(['icsevent'], { calendarId: calendar_id.to_s, eventId: event_id.to_s, ics: ics.to_s }.merge(options))
     end
 
     def delete_calendar(calendar_id)
@@ -83,8 +95,6 @@ module Teamlab
     def delete_event_series(event_id)
       @request.delete(['events', event_id.to_s])
     end
-
-    #============== TODO: OPTIONAL VARIABLES
 
     def remove_event(event_id, options = {})
       @request.delete(['events', event_id.to_s, 'custom'], options)